move fwdpy11.ezparams to tests/
This module was undocumented and only used
in the test suite.  Moving it is a breaking
change that is unlikely to actually break
any user scripts.
