+++ tests/quick_pops.py
@@ -21,9 +21,10 @@
 # This file contains functions for quickly
 # obtaining simulated populations
 
+from ezparams import mslike
+
 
 def quick_neutral_slocus(N=1000, simlen=100):
-    from fwdpy11.ezparams import mslike
     from fwdpy11 import ModelParams
     from fwdpy11 import DiploidPopulation, GSLrng
     from fwdpy11 import Multiplicative
@@ -39,7 +40,6 @@ def quick_neutral_slocus(N=1000, simlen=100):
 
 
 def quick_nonneutral_slocus(N=1000, simlen=100, dfe=None):
-    from fwdpy11.ezparams import mslike
     from fwdpy11 import ModelParams
     from fwdpy11 import DiploidPopulation, GSLrng
     from fwdpy11 import evolvets
